# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cat Breed Learning API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/quiz/sessions": {
      "post": {
        "tags": [
          "quiz"
        ],
        "summary": "Create Quiz Session",
        "description": "クイズセッションを作成し、10問を一括生成する。",
        "operationId": "create_quiz_session_api_v1_quiz_sessions_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quiz/today": {
      "get": {
        "tags": [
          "quiz"
        ],
        "summary": "Get Today Quiz",
        "description": "今日の一匹クイズを取得する。",
        "operationId": "get_today_quiz_api_v1_quiz_today_get",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayQuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quiz/answer": {
      "post": {
        "tags": [
          "quiz"
        ],
        "summary": "Submit Answer",
        "description": "クイズ回答を送信する。正解判定はサーバー側で行う。",
        "operationId": "submit_answer_api_v1_quiz_answer_post",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quiz/sessions/{session_id}/finalize": {
      "post": {
        "tags": [
          "quiz"
        ],
        "summary": "Finalize Session",
        "description": "セッションを完了し、結果をサーバー側で算出する。",
        "operationId": "finalize_session_api_v1_quiz_sessions__session_id__finalize_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          },
          {
            "name": "csrf_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Csrf Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/stats": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User Stats",
        "description": "ユーザー統計（累計覚えた種類数）を取得する。",
        "operationId": "get_user_stats_api_v1_users_me_stats_get",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/sessions/latest": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Latest Session",
        "description": "最新セッション結果を取得する。",
        "operationId": "get_latest_session_api_v1_users_me_sessions_latest_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "セッション種別 (quiz / today)",
              "title": "Source"
            },
            "description": "セッション種別 (quiz / today)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "question_number": {
            "type": "integer",
            "title": "Question Number"
          },
          "selected_cat_id": {
            "type": "string",
            "format": "uuid",
            "title": "Selected Cat Id"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "question_number",
          "selected_cat_id"
        ],
        "title": "AnswerRequest",
        "description": "POST /quiz/answer のリクエスト"
      },
      "AnswerResponse": {
        "properties": {
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "correct_cat_id": {
            "type": "string",
            "format": "uuid",
            "title": "Correct Cat Id"
          }
        },
        "type": "object",
        "required": [
          "is_correct",
          "correct_cat_id"
        ],
        "title": "AnswerResponse",
        "description": "POST /quiz/answer のレスポンス"
      },
...
      "FinalizeResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "correct_count": {
            "type": "integer",
            "title": "Correct Count"
          },
          "incorrect_count": {
            "type": "integer",
            "title": "Incorrect Count"
          },
          "correct_rate": {
            "type": "number",
            "title": "Correct Rate"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "source",
          "correct_count",
          "incorrect_count",
          "correct_rate",
          "completed_at"
        ],
        "title": "FinalizeResponse",
        "description": "POST /quiz/sessions/{session_id}/finalize のレスポンス"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LatestSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "correct_count": {
            "type": "integer",
            "title": "Correct Count"
          },
          "incorrect_count": {
            "type": "integer",
            "title": "Incorrect Count"
          },
          "correct_rate": {
            "type": "number",
            "title": "Correct Rate"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "source",
          "correct_count",
          "incorrect_count",
          "correct_rate",
          "completed_at"
        ],
        "title": "LatestSessionResponse",
        "description": "最新セッション結果レスポンス。\n\nsession_results を user_id + source で最新1件取得した結果。"
      },
...
      "QuizChoiceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "QuizChoiceResponse",
        "description": "選択肢の統一レスポンス。\n\nphoto_to_name: id + name が設定\nname_to_photo: id + photo_url が設定"
      },
      "QuizQuestionResponse": {
        "properties": {
          "question_number": {
            "type": "integer",
            "title": "Question Number"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "cat_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cat Name"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/QuizChoiceResponse"
            },
            "type": "array",
            "title": "Choices"
          }
        },
        "type": "object",
        "required": [
          "question_number",
          "question_type",
          "choices"
        ],
        "title": "QuizQuestionResponse",
        "description": "問題1件のレスポンス。\n\nquestion_type に応じて photo_url / cat_name / choices の中身が変わる。\n- photo_to_name: photo_url あり, choices は name のみ\n- name_to_photo: cat_name あり, choices は photo_url のみ"
      },
      "QuizSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionResponse"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "questions"
        ],
        "title": "QuizSessionResponse",
        "description": "POST /quiz/sessions のレスポンス"
      },
...
      "TodayQuizResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type"
          },
          "question_number": {
            "type": "integer",
            "title": "Question Number",
            "default": 1
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "cat_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cat Name"
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/QuizChoiceResponse"
            },
            "type": "array",
            "title": "Choices"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "question_type",
          "choices"
        ],
        "title": "TodayQuizResponse",
        "description": "GET /quiz/today のレスポンス（1問のみ）"
      },
...
      "UserStatsResponse": {
        "properties": {
          "total_correct_breeds": {
            "type": "integer",
            "title": "Total Correct Breeds"
          }
        },
        "type": "object",
        "required": [
          "total_correct_breeds"
        ],
        "title": "UserStatsResponse",
        "description": "ユーザー統計レスポンス。\n\ncorrect_answers テーブルの件数（UNIQUE制約により自動でユニーク数）。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}